moved functionality for converting time into js file in lib to help clean up code
